17th commit
Fully functional Pause menu created. During the active state of the Pause menu, every element and component in the game goes to 'standby mode'. If Player presses the Esc key during the gameplay, the Pause menu is activated. It contains two options, one returns the game to the Main menu and the second one exits the game completely. If the Esc key is pressed again, Pause menu terminates and Player continues with the standard gameplay.